# Use Spans and ReadOnlyCollection.Empty in TextBlock/Line, reduce allocs