--- a/src/Microsoft.DotNet.Wpf/src/PresentationFramework/System/Windows/Controls/TextBlock.cs
+++ b/src/Microsoft.DotNet.Wpf/src/PresentationFramework/System/Windows/Controls/TextBlock.cs
@@ -3449,10 +3449,7 @@ Debug.Assert(lineCount == LineCount);
         // ------------------------------------------------------------------
         private static void OnRequestBringIntoView(object sender, RequestBringIntoViewEventArgs args)
         {
-            TextBlock textBlock = sender as TextBlock;
-            ContentElement child = args.TargetObject as ContentElement;
-
-            if (textBlock != null && child != null)
+            if (sender is TextBlock textBlock && args.TargetObject is ContentElement child)
             {
                 if (TextBlock.ContainsContentElement(textBlock, child))
                 {